+++ pages/gestione/manutenzione/old_chat.inc.php
@@ -1,11 +1,19 @@
 <?php
+
+/*HELP: */
+/*Controllo permessi utente*/
+if(!gdrcd_controllo_permessi($PARAMETERS['administration']['maintenance']['access_level'])) {
+    echo '<div class="error">'.gdrcd_filter('out', $MESSAGE['error']['not_allowed']).'</div>';
+    die();
+}
+
 if((is_numeric($_POST['mesi']) === true) && ($_POST['mesi'] >= 0) && ($_POST['mesi'] <= 12)) {
     /*Eseguo l'aggiornamento*/
     gdrcd_query("DELETE FROM chat WHERE DATE_SUB(NOW(), INTERVAL ".gdrcd_filter('num', $_POST['mesi'])." MONTH) > ora");
     gdrcd_query("OPTIMIZE TABLE chat");
     ?>
     <!-- Conferma -->
-    <div class="warning">
+    <div class="success">
         <?php echo gdrcd_filter('out', $MESSAGE['warning']['modified']); ?>
     </div>
     <?php
@@ -16,10 +24,3 @@ if((is_numeric($_POST['mesi']) === true) && ($_POST['mesi'] >= 0) && ($_POST['me
     </div>
     <?php
 }
-?>
-<!-- Link di ritorno alla visualizzazione di base -->
-<div class="link_back">
-    <a href="main.php?page=gestione_manutenzione">
-        <?php echo gdrcd_filter('out', $MESSAGE['interface']['administration']['maintenance']['link']['back']); ?>
-    </a>
-</div>
